% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\docType{data}
\name{simul}
\alias{simul}
\title{Toy multilevel categorical dataset (simul)}
\format{A data frame with 13 columns and 1,000 rows (level-1 units) from 200 groups (level-2 units):
\tabular{llll}{
    [,1] \tab GroupID     \tab numeric  \tab Level-2 unit Identifier \cr
    [,2] \tab UnitID   \tab numeric  \tab Level-1 unit Identifier \cr
    [,3] \tab X1,...,X5 \tab binary  \tab Level-1 predictors \cr
    [,4] \tab Z1,...,Z5 \tab binary  \tab Level-2 predictors \cr
    [,5] \tab Y   \tab binary  \tab Response Variable\cr
}}
\usage{
data(simul)
}
\description{
The dataset contains 1000 level-1 artificial observations grouped into 200 level-2 groups. The dataset consists of: one identifier for the group, 
one identifier for the level-1 units, six level-1 binary variables (5 predictors and 1 response), five level-2 binary variables. The dataset was generated following the 
simulation study settings in Vidotto, Vermunt, van Deun (2018). The dataset can be used for performance evaluation of multilevel imputation methods applied on the 
\code{\link{simul_incomplete}} dataset.
}
\examples{
\dontrun{

library(BMLCimpute)

data(simul)

}


}
\references{
\itemize{

\item [1] Vidotto D., Vermunt J.K., Van Deun K. (2018). 'Bayesian Multilevel Latent Class Models for the Multiple Imputation of Nested Categorical Data'. Journal 
of Educational and Beahvioral Statistics 43(5), 511-539. 

}
}
\keyword{datasets}

